+++ DatadogTrace/Sources/TraceConfiguration.swift
@@ -19,6 +19,7 @@ import DatadogInternal
 @_exported import class DatadogInternal.W3CHTTPHeadersWriter
 @_exported import enum DatadogInternal.TraceSamplingStrategy
 @_exported import enum DatadogInternal.TraceContextInjection
+@_exported import enum DatadogInternal.TracingHeaderType
 // swiftlint:enable duplicate_imports
 
 extension Trace {